Apply improved types & architecture to Casper
