+++ libs/ledger-live-common/src/families/casper/estimateMaxSpendable.ts
@@ -1,0 +1,26 @@
+import BigNumber from "bignumber.js";
+import { AccountBridge } from "@ledgerhq/types-live";
+import { getMainAccount } from "@ledgerhq/coin-framework/account/index";
+import { getEstimatedFees } from "./bridge/bridgeHelpers/fee";
+import { Transaction } from "./types";
+
+export const estimateMaxSpendable: AccountBridge<Transaction>["estimateMaxSpendable"] = async ({
+  account,
+  parentAccount,
+  transaction,
+}) => {
+  const a = getMainAccount(account, parentAccount);
+  let balance = a.spendableBalance;
+
+  if (balance.eq(0)) return balance;
+
+  const estimatedFees = transaction?.fees ?? getEstimatedFees();
+
+  if (balance.lte(estimatedFees)) return new BigNumber(0);
+
+  balance = balance.minus(estimatedFees);
+
+  // log("debug", "[estimateMaxSpendable] finish fn");
+
+  return balance;
+};
